requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
bandit==1.6.2
black==19.10b0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-dotenv==1.4.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
dodgy==0.2.1
entrypoints==0.3
flake8==3.7.9
gitdb2==2.0.6
GitPython==3.0.5
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.761
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.7.0
pbr==5.4.4
pep8-naming==0.4.1
prospector==1.2.0
psycopg2-binary==2.8.4
pycodestyle==2.4.0
pycparser==2.19
pydocstyle==5.0.2
pyflakes==2.1.1
pylama==7.7.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.12
pylint-flask==0.6
pylint-plugin-utils==0.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
requirements-detector==0.6
setoptconf==0.2.0
six==1.14.0
smmap2==2.0.5
snowballstemmer==2.0.0
sqlparse==0.3.0
stevedore==1.31.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wrapt==1.11.2
yapf==0.29.0
